feat(uninstall): spinner during the deep-sweep daemon reload; share the spinner
The deep-sweep (`d`) reload starts the daemon, which on a cold cache can take
~90s. In loud mode that showed only raw "connect attempt N/20" chatter, which
reads as a hang. Animate a spinner over the (quiet) start instead so the wait
reads as "working".

Extract the spinner into a shared `commands::spinner` (Windows-only for now,
since both callers are `#[cfg(windows)]`) and point both the winget
orchestration and the new coverage reload at it, removing winget's private
copy (behavior-identical output).

native + windows-msvc clippy + rustdoc clean; 147 uffs-cli tests pass.
